Add dropdown menus to timepicker
Using sliders to change hours, minutes and seconds is cumbersome to many
users. Especially entering "clean" times like 09:15:00 is a pain, as the
sliders are pretty short.

This pull request adds dropdown menus to the hours, minutes and seconds
values. The "hours" dropdown menu has submenus for each hour with
00:00, 15:00, 30:00 and 45:00 settings for minutes and seconds. This
makes it easy to select clean time slots. The dropdown menus for
minutes and seconds menus have their values in 5 minutes/seconds steps,
with a submenu to select the values in between.

To minimize the required mouse interaction, the dropdowns and submenus
open on mouse over (hover). The min/max values for hours, minutes and
seconds are respected.

Also includes a trivial fix for the spacing between the minutes and
seconds slider.